+++ src/dapp/walletProvider.ts
@@ -10,19 +10,13 @@ import {
 } from "./constants";
 import { Transaction } from "../transaction";
 import { SignableMessage } from "../signableMessage";
-import {ErrInvalidTxSignReturnValue, ErrNotImplemented} from "../errors";
-import {
-    Nonce,
-    Balance,
-    Address,
-    GasPrice,
-    GasLimit,
-    TransactionPayload,
-    ChainID,
-    TransactionVersion,
-    TransactionOptions
-} from "../";
-import {Signature} from "../signature";
+import { ErrInvalidTxSignReturnValue, ErrNotImplemented } from "../errors";
+import { Signature } from "../signature";
+import { Nonce } from "../nonce";
+import { Balance } from "../balance";
+import { Address } from "../address";
+import { ChainID, GasLimit, GasPrice, TransactionOptions, TransactionVersion } from "../networkParams";
+import { TransactionPayload } from "../transactionPayload";
 
 interface TransactionMessage {
     receiver: string;
@@ -256,7 +250,7 @@ export class WalletProvider implements IDappProvider {
         if (transaction.data) {
             urlString += `&data=${transaction.data}`;
         }
-        if (transaction.nonce) {
+        if (transaction.nonce || transaction.nonce === 0) {
             urlString += `&nonce=${transaction.nonce}`;
         }