+++ src/smartcontracts/codec/primitive.ts
@@ -7,6 +7,7 @@ import {
     BooleanValue,
     NumericalValue,
     AddressValue,
+    StringValue,
 } from "../typesystem";
 import { AddressBinaryCodec } from "./address";
 import { BooleanBinaryCodec } from "./boolean";
@@ -19,6 +20,7 @@ import { BytesBinaryCodec } from "./bytes";
 import { TokenIdentifierCodec } from "./tokenIdentifier";
 import { TokenIdentifierValue } from "../typesystem/tokenIdentifier";
 import { NothingCodec } from "./nothing";
+import { StringBinaryCodec } from "./string";
 
 export class PrimitiveBinaryCodec {
     private readonly binaryCodec: BinaryCodec;
@@ -28,6 +30,7 @@ export class PrimitiveBinaryCodec {
     private readonly addressCodec: AddressBinaryCodec;
     private readonly h256Codec: H256BinaryCodec;
     private readonly bytesCodec: BytesBinaryCodec;
+    private readonly stringCodec: StringBinaryCodec;
     private readonly tokenIdentifierCodec: TokenIdentifierCodec;
     private readonly nothingCodec: NothingCodec;
 
@@ -39,6 +42,7 @@ export class PrimitiveBinaryCodec {
         this.addressCodec = new AddressBinaryCodec();
         this.h256Codec = new H256BinaryCodec();
         this.bytesCodec = new BytesBinaryCodec();
+        this.stringCodec = new StringBinaryCodec();
         this.tokenIdentifierCodec = new TokenIdentifierCodec();
         this.nothingCodec = new NothingCodec();
     }
@@ -49,6 +53,7 @@ export class PrimitiveBinaryCodec {
             onNumerical: () => this.numericalCodec.decodeNested(buffer, <NumericalType>type),
             onAddress: () => this.addressCodec.decodeNested(buffer),
             onBytes: () => this.bytesCodec.decodeNested(buffer),
+            onString: () => this.stringCodec.decodeNested(buffer),
             onH256: () => this.h256Codec.decodeNested(buffer),
             onTokenIndetifier: () => this.tokenIdentifierCodec.decodeNested(buffer),
             onNothing: () => this.nothingCodec.decodeNested()
@@ -61,6 +66,7 @@ export class PrimitiveBinaryCodec {
             onNumerical: () => this.numericalCodec.decodeTopLevel(buffer, <NumericalType>type),
             onAddress: () => this.addressCodec.decodeTopLevel(buffer),
             onBytes: () => this.bytesCodec.decodeTopLevel(buffer),
+            onString: () => this.stringCodec.decodeTopLevel(buffer),
             onH256: () => this.h256Codec.decodeTopLevel(buffer),
             onTokenIndetifier: () => this.tokenIdentifierCodec.decodeTopLevel(buffer),
             onNothing: () => this.nothingCodec.decodeTopLevel()
@@ -73,6 +79,7 @@ export class PrimitiveBinaryCodec {
             onNumerical: () => this.numericalCodec.encodeNested(<NumericalValue>value),
             onAddress: () => this.addressCodec.encodeNested(<AddressValue>value),
             onBytes: () => this.bytesCodec.encodeNested(<BytesValue>value),
+            onString: () => this.stringCodec.encodeNested(<StringValue>value),
             onH256: () => this.h256Codec.encodeNested(<H256Value>value),
             onTypeIdentifier: () => this.tokenIdentifierCodec.encodeNested(<TokenIdentifierValue>value),
             onNothing: () => this.nothingCodec.encodeNested()
@@ -85,6 +92,7 @@ export class PrimitiveBinaryCodec {
             onNumerical: () => this.numericalCodec.encodeTopLevel(<NumericalValue>value),
             onAddress: () => this.addressCodec.encodeTopLevel(<AddressValue>value),
             onBytes: () => this.bytesCodec.encodeTopLevel(<BytesValue>value),
+            onString: () => this.stringCodec.encodeTopLevel(<StringValue>value),
             onH256: () => this.h256Codec.encodeTopLevel(<H256Value>value),
             onTypeIdentifier: () => this.tokenIdentifierCodec.encodeTopLevel(<TokenIdentifierValue>value),
             onNothing: () => this.nothingCodec.encodeTopLevel()