+++ src/smartcontracts/typesystem/matchers.ts
@@ -12,6 +12,8 @@ import { TokenIdentifierType, TokenIdentifierValue } from "./tokenIdentifier";
 import { Tuple, TupleType } from "./tuple";
 import { Type, PrimitiveType, PrimitiveValue } from "./types";
 import { ArrayVec, ArrayVecType } from "./genericArray";
+import { TypedValue } from "./types";
+import { StringType, StringValue } from "./string";
 
 // TODO: Extend functionality or rename wrt. restricted / reduced functionality (not all types are handled: composite, variadic).
 export function onTypeSelect<TResult>(
@@ -27,25 +29,25 @@ export function onTypeSelect<TResult>(
         onOther?: () => TResult;
     }
 ): TResult {
-    if (type instanceof OptionType) {
+    if (type.hasJavascriptConstructorInHierarchy(OptionType.name)) {
         return selectors.onOption();
     }
-    if (type instanceof ListType) {
+    if (type.hasJavascriptConstructorInHierarchy(ListType.name)) {
         return selectors.onList();
     }
-    if (type instanceof ArrayVecType) {
+    if (type.hasJavascriptConstructorInHierarchy(ArrayVecType.name)) {
         return selectors.onArray();
     }
-    if (type instanceof PrimitiveType) {
+    if (type.hasJavascriptConstructorInHierarchy(PrimitiveType.name)) {
         return selectors.onPrimitive();
     }
-    if (type instanceof StructType) {
+    if (type.hasJavascriptConstructorInHierarchy(StructType.name)) {
         return selectors.onStruct();
     }
-    if (type instanceof TupleType) {
+    if (type.hasJavascriptConstructorInHierarchy(TupleType.name)) {
         return selectors.onTuple();
     }
-    if (type instanceof EnumType) {
+    if (type.hasJavascriptConstructorInHierarchy(EnumType.name)) {
         return selectors.onEnum();
     }
 
@@ -57,7 +59,7 @@ export function onTypeSelect<TResult>(
 }
 
 export function onTypedValueSelect<TResult>(
-    value: any,
+    value: TypedValue,
     selectors: {
         onPrimitive: () => TResult;
         onOption: () => TResult;
@@ -69,25 +71,25 @@ export function onTypedValueSelect<TResult>(
         onOther?: () => TResult;
     }
 ): TResult {
-    if (value instanceof PrimitiveValue) {
+    if (value.hasJavascriptConstructorInHierarchy(PrimitiveValue.name)) {
         return selectors.onPrimitive();
     }
-    if (value instanceof OptionValue) {
+    if (value.hasJavascriptConstructorInHierarchy(OptionValue.name)) {
         return selectors.onOption();
     }
-    if (value instanceof List) {
+    if (value.hasJavascriptConstructorInHierarchy(List.name)) {
         return selectors.onList();
     }
-    if (value instanceof ArrayVec) {
+    if (value.hasJavascriptConstructorInHierarchy(ArrayVec.name)) {
         return selectors.onArray();
     }
-    if (value instanceof Struct) {
+    if (value.hasJavascriptConstructorInHierarchy(Struct.name)) {
         return selectors.onStruct();
     }
-    if (value instanceof Tuple) {
+    if (value.hasJavascriptConstructorInHierarchy(Tuple.name)) {
         return selectors.onTuple();
     }
-    if (value instanceof EnumValue) {
+    if (value.hasJavascriptConstructorInHierarchy(EnumValue.name)) {
         return selectors.onEnum();
     }
 
@@ -105,31 +107,35 @@ export function onPrimitiveValueSelect<TResult>(
         onNumerical: () => TResult;
         onAddress: () => TResult;
         onBytes: () => TResult;
+        onString: () => TResult;
         onH256: () => TResult;
         onTypeIdentifier: () => TResult;
         onNothing: () => TResult;
         onOther?: () => TResult;
     }
 ): TResult {
-    if (value instanceof BooleanValue) {
+    if (value.hasJavascriptConstructorInHierarchy(BooleanValue.name)) {
         return selectors.onBoolean();
     }
-    if (value instanceof NumericalValue) {
+    if (value.hasJavascriptConstructorInHierarchy(NumericalValue.name)) {
         return selectors.onNumerical();
     }
-    if (value instanceof AddressValue) {
+    if (value.hasJavascriptConstructorInHierarchy(AddressValue.name)) {
         return selectors.onAddress();
     }
-    if (value instanceof BytesValue) {
+    if (value.hasJavascriptConstructorInHierarchy(BytesValue.name)) {
         return selectors.onBytes();
     }
-    if (value instanceof H256Value) {
+    if (value.hasJavascriptConstructorInHierarchy(StringValue.name)) {
+        return selectors.onString();
+    }
+    if (value.hasJavascriptConstructorInHierarchy(H256Value.name)) {
         return selectors.onH256();
     }
-    if (value instanceof TokenIdentifierValue) {
+    if (value.hasJavascriptConstructorInHierarchy(TokenIdentifierValue.name)) {
         return selectors.onTypeIdentifier();
     }
-    if (value instanceof NothingValue) {
+    if (value.hasJavascriptConstructorInHierarchy(NothingValue.name)) {
         return selectors.onNothing();
     }
     if (selectors.onOther) {
@@ -146,31 +152,35 @@ export function onPrimitiveTypeSelect<TResult>(
         onNumerical: () => TResult;
         onAddress: () => TResult;
         onBytes: () => TResult;
+        onString: () => TResult;
         onH256: () => TResult;
         onTokenIndetifier: () => TResult;
         onNothing: () => TResult;
         onOther?: () => TResult;
     }
 ): TResult {
-    if (type instanceof BooleanType) {
+    if (type.hasJavascriptConstructorInHierarchy(BooleanType.name)) {
         return selectors.onBoolean();
     }
-    if (type instanceof NumericalType) {
+    if (type.hasJavascriptConstructorInHierarchy(NumericalType.name)) {
         return selectors.onNumerical();
     }
-    if (type instanceof AddressType) {
+    if (type.hasJavascriptConstructorInHierarchy(AddressType.name)) {
         return selectors.onAddress();
     }
-    if (type instanceof BytesType) {
+    if (type.hasJavascriptConstructorInHierarchy(BytesType.name)) {
         return selectors.onBytes();
     }
-    if (type instanceof H256Type) {
+    if (type.hasJavascriptConstructorInHierarchy(StringType.name)) {
+        return selectors.onString();
+    }
+    if (type.hasJavascriptConstructorInHierarchy(H256Type.name)) {
         return selectors.onH256();
     }
-    if (type instanceof TokenIdentifierType) {
+    if (type.hasJavascriptConstructorInHierarchy(TokenIdentifierType.name)) {
         return selectors.onTokenIndetifier();
     }
-    if (type instanceof NothingType) {
+    if (type.hasJavascriptConstructorInHierarchy(NothingType.name)) {
         return selectors.onNothing();
     }
     if (selectors.onOther) {