+++ src/smartcontracts/wrapper/contractWrapper.local.net.spec.ts
@@ -1,7 +1,9 @@
-import { SystemWrapper, Balance, setupInteractive } from "../..";
 import { assert } from "chai";
 import { BigNumber } from "bignumber.js";
 import { TestWallet } from "../../testutils";
+import { SystemWrapper } from "./systemWrapper";
+import { setupInteractive } from "../../interactive";
+import { Balance } from "../../balance";
 
 
 describe("test smart contract interactor", function () {
@@ -53,7 +55,7 @@ describe("test smart contract interactor", function () {
         await lottery.call.start("lucky", Balance.egld(1), null, null, 1, null, null);
 
         let status = await lottery.query.status("lucky");
-        assert.equal(status.valueOf(), "Running");
+        assert.equal(status.valueOf().name, "Running");
 
         let info = await lottery.query.lotteryInfo("lucky");
         // Ignore "deadline" field in our test